Add RepositoryTest, PollControllerTest + Refactoring.

+++ src/main/java/dat250/msd/FeedApp/controller/PollController.java
@@ -1,5 +1,6 @@
 package dat250.msd.FeedApp.controller;
 
+import dat250.msd.FeedApp.model.Instance;
 import dat250.msd.FeedApp.model.Poll;
 import dat250.msd.FeedApp.model.UserData;
 import dat250.msd.FeedApp.model.VoteOption;
@@ -23,21 +24,38 @@ public class PollController {
 
     /**
      * Create a new Poll. A poll should at least have a name and owner.
-     * Vote options added with a PUT operation and Instances can be connected trough /instance POST.
+     * Vote options added with a PUT operation and Instances can be connected trough /instance POST:
+     * Example of POST /poll?username=Testuser1&pwd=123
+     * {
+     *     "name": "Test Poll",
+     *     "voteOptions": [
+     *         {
+     *             "label": "Yes"
+     *         },
+     *         {
+     *             "label": "Maybe"
+     *         },
+     *         {
+     *             "label": "No"
+     *         }
+     *     ]
+     * }
      * */
     @PostMapping("/poll")
-    public Poll createPoll(@RequestParam String username, @RequestParam String pwd, @RequestParam String pollName){
+    public Poll createPoll(@RequestParam String username, @RequestParam String pwd, @RequestBody Poll poll){
         UserData owner = feedAppService.getUser(username,pwd);
-
-        Poll poll = new Poll();
+        if (owner == null || poll.getName() == null){
+            //TODO error
+            return new Poll();
+        }
         poll.setOwner(owner);
-        poll.setName(pollName);
 
         return feedAppService.getPollRepository().save(poll);
     }
 
     /**
      * Update the poll voteOptions using the poll id and a json body like:
+     * PUT: poll?id=1
      * [
      *     {
      *         "label": "Frozen Toast"
@@ -65,6 +83,16 @@ public class PollController {
     public Poll deletePoll(@RequestParam Long id){
         //TODO require auth
         Poll poll = feedAppService.getPollRepository().getPollById(id);
+        if (poll == null){
+            System.out.println("No poll with id: "+id);
+            return new Poll();
+        }
+
+        // Remove all votes from every instance
+        for (Instance instance : poll.getInstances()){
+            feedAppService.removeVotes(instance);
+        }
+
         feedAppService.getPollRepository().delete(poll);
         return poll;
     }
